Added code for evaluation

+++ Code/utils/utils.h
@@ -7,6 +7,6 @@ double similarity(std::vector<std::string> a, std::vector<std::string> b);
 
 std::map<std::string, double> distances(std::vector<std::vector<std::string> > a);
 
-std::vector<std::size_t> clustering(std::vector<std::vector<std::string> > a, int max_d);
+std::vector<std::size_t> clustering(std::vector<std::vector<std::string> > a, double max_d);
 
-std::vector<std::size_t> clustering_lsh(std::vector<std::vector<std::string> > a, std::set<std::string> pairs, int max_d);
+std::vector<std::size_t> clustering_lsh(std::vector<std::vector<std::string> > a, std::set<std::string> pairs, double max_d);
